Remove FreeRTOS submodule for now.
It turns out that we cannot have FreeRTOS submodule in the same git repo
as the crazyflie firmware (since the make command tries to compile stuff in
FreeRTOS submodule as well, and that throws errors).

In order to have FreeRTOS for the course, we should put it in another repo
(or, say, a parent repo) where is does not conflict with the crazyflie firmware.